requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.2.3
django-appconf==1.0.4
django-compressor==2.4
django-filter==2.2.0
django-haystack==2.8.1
django-mdeditor==0.1.18
django-pure-pagination==0.3.0
djangorestframework==3.11.0
docopt==0.6.2
faker
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
invoke==1.4.1
jsonschema==3.2.0
Markdown==3.2.2
mistune==0.8.4
paramiko==2.7.1
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
texttable==1.6.2
urllib3==1.25.9
websocket-client==0.57.0
zipp==3.1.0
dj-database-url==0.5.0
dj-static==0.0.6
django-heroku==0.3.1
psycopg2==2.8.4
static3==0.7.0
whitenoise==5.0.1
